fix: flah + hm11 init fault

+++ ODB_V3/Software/STM32CubeIDE/ODBlix/Core/Src/GAUL_Drivers/w25q512jv.c
@@ -29,7 +29,7 @@ static QSPI_CommandTypeDef W25Q_MakeCommand(uint32_t instruction, uint32_t addre
     return sCommand;
 }
 
-static uint8_t W25Q_GetID(QSPI_HandleTypeDef *hqspi) {
+static uint32_t W25Q_GetID(QSPI_HandleTypeDef *hqspi) {
     QSPI_CommandTypeDef sCommand = W25Q_MakeCommand(W25Q_CMD_READ_ID, QSPI_ADDRESS_NONE, 0xFFFFFFFF, QSPI_DATA_1_LINE, 0, 3);
     if(HAL_QSPI_Command(hqspi, &sCommand, HAL_QSPI_TIMEOUT_DEFAULT_VALUE) != HAL_OK) return -1;
